requirements.txt
mridata
tensorflow-gpu==1.12
numpy
scipy
ismrmrd